component improvements
- Remove animations on every item of the list
- Implement gradient to achieve similar effect
- Break down model, and the selectedOption component into 2 separate modules
- Expose additional prop on the modal list, allowing you to use custom render method for every row
- Export non-decorated base classes
